Use new mocks package to simplify build and tests
Created a new repo that contains the TypeScript typings as well as a few
mocks. Using triple slash directive, the compiler can be directed to the
package to find the typings it needs to be able to build. The package
also contains mocks and util classes so that the Karma tests work too.