wip integrating the monaco editor
This includes a custom built monaco js file to get
around problems with the AMD format. Grafana uses SystemJS
to import plugins and those plugins must be in the systemjs
or AMD formats. This causes problems for Monaco as it uses web
workers to dynamically load code and they cannot automatically
load AMD code. The solution was to create one big bundle and
wrap that in a define function.

This bundle includes everything so it has to be trimmed down
to a reasonable size before release.

This commit just includes it in the query ctrl so the html and js
there is just test code to get the monaco editor working. This will
be removed and done properly in a later commit