Fixes #27. Support long metric definitions.
It seems if a metric definition contains 2 or more forward slashes then
it should be split before making a request back to Azure. Not sure if this is
how it should be or just inconsistency with different parts of Azure.

The exception to this rule is Microsoft.Compute/virtualMachines/extensions which
does not seem work no matter how the query url is formed.