azureloganalytics: http cache requests for 1 minute
To prevent the Azure Log Analytics service being overloaded,
this change throttles multiple identical requests using http
caching (the Cache-Control header). To make sure that refresh
settings do not send a lot of unique queries, the from and to
times in the selected date range are rounded to the nearest
minute. Previously the from and to fields had millisecond precision.